rename dpctl.dptensor -> dpctl.tensor
