+++ dpctl/tensor/numpy_usm_shared.py
